safari-600.7-branch/LayoutTests/platform/ios-sim/fast/workers/worker-copy-shared-blob-url-expected.txt

CONSOLE MESSAGE: line 53: TypeError: undefined is not a function (evaluating 'eventSender.beginDragWithFiles(["worker-copy-shared-blob-urls.html"])')
FAIL: Timed out waiting for notifyDone to be called

Test result of transferring a file and filename between worker and main page. Should print a series of PASS messages, followed with DONE.
